requirements.txt
pytest==8.1.1
jupyter==1.0.0
numpy==1.24.1
pandas==2.0.2 
pytz==2022.6
psycopg2-binary>=2.8.6
matplotlib>=3.7.1
ipywidgets>=8.0.0
ipython>=8.0.0